backend/requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
python-jose[cryptography]==3.2.0
pydantic==1.8.1
passlib[bcrypt]==1.7.4
psycopg2==2.8.6
python-multipart==0.0.5